refactor: remove old IO code
