+++ src/io/io.ts
@@ -1,32 +1,34 @@
-import { vtkObject } from '@kitware/vtk.js/interfaces';
+import extensionToImageIO from 'itk/extensionToImageIO';
+import { extractFilesFromZip } from './zip';
 
-// How much data to read when extracting file magic
-const HEAD_CHUNK = 512;
+export const ARCHIVE_FILE_TYPES = new Set(['zip']);
 
-export type ReadAsType = 'ArrayBuffer' | 'Text';
+export const ITK_IMAGE_EXTENSIONS = Array.from(
+  new Set(Array.from(extensionToImageIO.keys()).map((ext) => ext.toLowerCase()))
+);
 
-export type ReaderFunction = (file: File) => vtkObject;
-
-/**
- * special file types that we handle specifically
- */
-export const FileTypes = {
-  DICOM: 'dcm',
-};
+// remove duplicates
+export const FILE_TYPES = Array.from(
+  new Set(['vti', 'vtp', 'stl', 'dcm', 'zip', ...ITK_IMAGE_EXTENSIONS])
+);
 
 /**
  * file magic database
  * Used to handle certain cases where files have no extension
  */
-export const FILE_MAGIC_DB = [
+const FILE_MAGIC_DB = [
   {
-    type: FileTypes.DICOM,
+    type: 'dcm',
     skip: 128,
     header: Array.from('DICM').map((c) => c.charCodeAt(0)),
   },
 ];
 
-function prefixEquals(target: Uint8Array, prefix: number[]) {
+// How much data to read when extracting file magic.
+// This should be generous enough for most files.
+const HEAD_CHUNK = 512;
+
+function prefixEquals(target: Uint8Array, prefix: number[] | Uint8Array) {
   if (prefix.length > target.length) {
     return false;
   }
@@ -38,13 +40,7 @@ function prefixEquals(target: Uint8Array, prefix: number[]) {
   return true;
 }
 
-/**
- * Returns file type based on magic
- * @async
- * @param {File} file
- * @returns {string|null}
- */
-export async function getFileMagic(file: File): Promise<string | null> {
+async function getFileTypeFromMagic(file: File): Promise<string | null> {
   return new Promise((resolve) => {
     const head = file.slice(0, HEAD_CHUNK);
     const reader = new FileReader();
@@ -62,6 +58,61 @@ export async function getFileMagic(file: File): Promise<string | null> {
     reader.readAsArrayBuffer(head);
   });
 }
+
+/**
+ * Retypes a given File.
+ *
+ * Handy for when a file object has no type given.
+ * Type is inferred from extension or magic.
+ *
+ * If type is not supported, file.type will be "".
+ * @param file
+ */
+export async function retypeFile(file: File): Promise<File> {
+  let type: string | null;
+
+  type =
+    FILE_TYPES.find((ext) => file.name.toLowerCase().endsWith(`.${ext}`)) ??
+    null;
+
+  if (!type) {
+    type = await getFileTypeFromMagic(file);
+  }
+
+  if (!type) {
+    type = '';
+  }
+
+  const retypedFile = new File([file], file.name, { type: type.toLowerCase() });
+  return retypedFile;
+}
+
+export async function extractArchivesRecursively(
+  files: File[]
+): Promise<File[]> {
+  const results = await Promise.all(
+    files.map(async (file) => {
+      if (file.type === 'zip') {
+        return extractFilesFromZip(file);
+      }
+      return [file];
+    })
+  );
+
+  const archivedFiles = await Promise.all(
+    results.flat().map((f) => retypeFile(f))
+  );
+  const archives = archivedFiles.filter(({ type }) =>
+    ARCHIVE_FILE_TYPES.has(type)
+  );
+  if (archives.length > 0) {
+    const moreFiles = await extractArchivesRecursively(archives);
+    return [...archivedFiles, ...moreFiles];
+  }
+  return archivedFiles;
+}
+
+export type ReadAsType = 'ArrayBuffer' | 'Text';
 
 async function readFileAs<T extends ArrayBuffer | string>(
   file: File,
@@ -96,136 +147,4 @@ export async function readFileAsArrayBuffer(file: File) {
  */
 export async function readFileAsUTF8Text(file: File) {
   return readFileAs<string>(file, 'Text');
-}
-
-export class FileIO {
-  fileReaders: Record<string, ReaderFunction>;
-
-  typeAliases: Record<string, string>;
-
-  typeCache: WeakMap<File, string>;
-
-  constructor() {
-    this.fileReaders = Object.create(null);
-    this.typeAliases = Object.create(null);
-    // Cache for file type. Prevents re-reading magic every time.
-    this.typeCache = new WeakMap();
-  }
-
-  /**
-   * readerFunc parses a given file.
-   * @callback readerFunc
-   * @param {File} file
-   * @returns {any}
-   */
-  /**
-   * Adds a reader for a file type.
-   *
-   * All file types are treated case-insensitively.
-   * @param {String} fileType
-   * @param {readerFunc} readerFunc
-   */
-  addSingleReader(fileType: string, readerFunc: ReaderFunction) {
-    this.fileReaders[fileType.toLowerCase()] = readerFunc;
-  }
-
-  /**
-   * Adds type aliases for a particular file type.
-   *
-   * All types are treated case-insensitively.
-   * The baseType must already be registered to a reader.
-   * Example: jpg is equivalent to jpeg
-   *
-   * @param {String} baseType
-   * @param {String[]} aliases
-   */
-  addFileTypeAliases(baseType: string, aliases: string[]) {
-    if (baseType.toLowerCase() in this.fileReaders) {
-      aliases.forEach((alias) => {
-        this.typeAliases[alias.toLowerCase()] = baseType;
-      });
-    }
-  }
-
-  /**
-   * Infers the file type from a File object
-   *
-   * @async
-   * @param {File} file
-   * @returns {String|null}
-   */
-  async getFileType(file: File): Promise<string | null> {
-    if (this.typeCache.has(file)) {
-      return this.typeCache.get(file) ?? null;
-    }
-
-    let type = null;
-
-    // first see if file matches a registered type
-    const registeredTypes = Object.keys(this.fileReaders);
-    for (let i = 0; i < registeredTypes.length; i += 1) {
-      if (file.name.toLowerCase().endsWith(`.${registeredTypes[i]}`)) {
-        type = registeredTypes[i];
-        break;
-      }
-    }
-
-    if (!type) {
-      // see if there's a type alias
-      const aliases = Object.keys(this.typeAliases);
-      for (let i = 0; i < aliases.length; i += 1) {
-        if (file.name.toLowerCase().endsWith(`.${aliases[i]}`)) {
-          type = this.typeAliases[aliases[i]];
-          break;
-        }
-      }
-    }
-
-    if (!type) {
-      // read file mimetype
-      const magic = await getFileMagic(file);
-      if (magic) {
-        type = magic.toLowerCase();
-      }
-    }
-
-    if (type) {
-      this.typeCache.set(file, type);
-    }
-    return type;
-  }
-
-  /**
-   * Determines if a file can be read.
-   *
-   * @async
-   * @param {File} file
-   * @returns {Boolean}
-   */
-  async canReadFile(file: File) {
-    const type = await this.getFileType(file);
-    return type ? !!this.fileReaders[type] : false;
-  }
-
-  /**
-   * Parses a single file to produce a single output dataset.
-   *
-   * @async
-   * @param {File} file
-   * @returns {any}
-   * @throws {Error}
-   */
-  async readSingleFile(file: File) {
-    const type = await this.getFileType(file);
-    if (!type) {
-      throw new Error(`No type info found for ${file.name}`);
-    }
-
-    const reader = this.fileReaders[type];
-    if (!reader) {
-      throw new Error(`No reader found for ${file.name}`);
-    }
-
-    return reader(file);
-  }
 }
