+++ src/main.js
@@ -13,13 +13,11 @@ import App from './components/App.vue';
 import createStore from './store';
 import vuetify from './plugins/vuetify';
 import { ProxyManagerVuePlugin } from './plugins/proxyManager';
-import { FileIO } from './io/io';
 import { DICOMIO } from './io/dicom';
-import { createTREReader, registerAllReaders } from './io/readers';
 import { setCurrentInstance } from './instances';
 import proxyConfiguration from './vtk/proxy';
 import WidgetProvider from './widgets/widgetProvider';
-import { FileIOInst, DICOMIOInst, ProxyManagerInst } from './constants';
+import { DICOMIOInst, ProxyManagerInst } from './constants';
 import { updateRulerFromWidgetStateEvent } from './store/tools/rulers';
 import ProxyManager from './core/proxies';
 import { provideToolManagers, CorePiniaProviderPlugin } from './core/provider';
@@ -35,17 +33,9 @@ Vue.use(PiniaVuePlugin);
 const proxyManager = vtkProxyManager.newInstance({ proxyConfiguration });
 setCurrentInstance(ProxyManagerInst, proxyManager);
 
-const fileIO = new FileIO();
-registerAllReaders(fileIO);
-setCurrentInstance(FileIOInst, fileIO);
-
 const dicomIO = new DICOMIO();
 dicomIO.initialize();
 setCurrentInstance(DICOMIOInst, dicomIO);
-
-// Right now, TRE reader depends on the DicomIO module since
-// that's where the TRE read logic resides.
-fileIO.addSingleReader('tre', createTREReader(dicomIO));
 
 // Initialize global mapper topologies
 // polys and lines in the front
@@ -58,7 +48,6 @@ vtkImageMapper.setResolveCoincidentTopologyPolygonOffsetParameters(1, 1);
 
 const dependencies = {
   proxyManager,
-  fileIO,
   dicomIO,
 };
